Fix => remaining amount type moved from float to decimal

+++ GivingBackChange/GivingBackChange.Business/Services/Implementations/GetChangeServices/GetChangeDynamicProgrammingService.cs
@@ -12,7 +12,7 @@ namespace GivingBackChange.Business.Services.Implementations.GetChangeServices
         {
         }
 
-        public override async Task<List<CoinBo>> GetChange(float remaningAmount)
+        public override async Task<List<CoinBo>> GetChange(decimal remaningAmount)
         {
             var coinReferential = (await this._coinBoxService.GetCoins()).OrderBy(c => c.Value).ToList();
             var remaningAmountInCents = this.GetRemainingAmountInCents(remaningAmount);
